Improve the pool by supporting instance invalidation
In some cases we know that a pool in the instance shouldn't be in use
anymore, we should then be able to mark that instance as invalid. So we
now have a `pool.invalidate()` for that. All instances are now added to
a cache which also includes the use count.